refactor: extract profile-config bootstrap into engine library
Slop-review #11 — moves the first-launch profile-config bootstrap
out of AppDelegate (which had no business owning TOML loading +
Hammerspoon migration + starter-config seeding) and into the engine
library where the rest of the Config layer lives.

New: Sources/AVPainReliever/Config/ProfileBootstrapper.swift

  public struct ProfileBootstrapper {
      public static let canonicalTOMLURL: URL
      public func loadOrBootstrap(logger: ApplierLogger) -> [Profile]
  }

The discovery / migration / seed logic is byte-identical to the prior
AppDelegate.loadProfiles — same priority order (canonical TOML →
Hammerspoon migration → starter config), same logging shape, same
fallback semantics. The starter TOML string is also unchanged.

AppDelegate changes:
- bootEngine() calls ProfileBootstrapper().loadOrBootstrap(logger:)
- The two remaining configURL references (Engine init + delete-profile
  ProfileWriter call) point at ProfileBootstrapper.canonicalTOMLURL
  instead of the now-deleted AppDelegate.profilesTOMLURL static
- Drops loadProfiles, starterConfig, writeStarterConfig (~120 lines)

AppDelegate: 748 → 628 LOC. The bootstrap behavior is now testable
without the app target.

No behavior changes. swift test: 203 passed before and after.
