+++ Sources/AVPainRelieverApp/AppDelegate.swift
@@ -187,12 +187,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
             .store(in: &cancellables)
     }
 
-    private static let profilesTOMLURL: URL =
-        URL(fileURLWithPath: NSHomeDirectory())
-            .appendingPathComponent(
-                "Library/Application Support/AVPainReliever/profiles.toml"
-            )
-
     /// The most recent profile name we surfaced through
     /// `onProfileApplied`. Used to suppress a notification for the
     /// initial evaluation on launch (the menu-bar title is already
@@ -342,7 +336,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
         engine?.stop()
 
         let logger = ConsoleLogger()
-        let profiles = loadProfiles(logger: logger)
+        let profiles = ProfileBootstrapper().loadOrBootstrap(logger: logger)
         availableProfiles = profiles
         let engine = buildEngine(profiles: profiles, logger: logger)
         engine.onProfileApplied = { [weak self] profile in
@@ -500,7 +494,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
             watcher: IOKitUSBWatcher(),
             audioController: CoreAudioController(),
             cameraController: AVFoundationCameraController(),
-            configURL: Self.profilesTOMLURL,
+            configURL: ProfileBootstrapper.canonicalTOMLURL,
             editing: editing,
             existingProfileSlugs: existing,
             virtualCameraEnabled: virtualCameraActivator.state == .on,
@@ -550,7 +544,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
     func deleteProfile(_ profile: Profile) {
         let pretty = PrettyName.format(profile.name)
         do {
-            try ProfileWriter().delete(named: profile.name, in: Self.profilesTOMLURL)
+            try ProfileWriter().delete(named: profile.name, in: ProfileBootstrapper.canonicalTOMLURL)
             reloadConfig()
         } catch {
             let failure = NSAlert()
@@ -631,118 +625,4 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
     /// Mirror of `settings.menuBarIconSymbol` for the same reason.
     var menuBarIconSymbol: String { settings.menuBarIconSymbol }
 
-    /// Profile-config discovery in priority order. Mirrors what the
-    /// eventual first-run wizard will codify, but for now lets a
-    /// developer or migrating Phase 1 user run the app immediately:
-    ///
-    ///   1. `~/Library/Application Support/AVPainReliever/profiles.toml`
-    ///      — the canonical Swift-app config location.
-    ///   2. `~/.hammerspoon/profiles.lua` — auto-import for users
-    ///      migrating from the Hammerspoon Phase 1 engine.
-    ///   3. **Bootstrap a default profiles.toml** — a fresh user
-    ///      with neither file gets a working "laptop" fallback
-    ///      profile written to (1) so the app actually does
-    ///      something on first launch.
-    private func loadProfiles(logger: ApplierLogger) -> [Profile] {
-        let tomlURL = Self.profilesTOMLURL
-        let luaURL = URL(fileURLWithPath: NSHomeDirectory())
-            .appendingPathComponent(".hammerspoon/profiles.lua")
-
-        if FileManager.default.fileExists(atPath: tomlURL.path) {
-            do {
-                let profiles = try ConfigLoader().loadProfiles(from: tomlURL)
-                logger.info("loaded \(profiles.count) profiles from \(tomlURL.path)")
-                return profiles
-            } catch {
-                logger.warn("failed to load \(tomlURL.path): \(error)")
-            }
-        }
-
-        if FileManager.default.fileExists(atPath: luaURL.path) {
-            do {
-                let lua = try String(contentsOf: luaURL, encoding: .utf8)
-                let importer = ConfigImporter()
-                let profiles = try importer.parse(lua)
-                // One-shot migration: write the imported profiles to
-                // the canonical TOML location so future launches AND
-                // the Add-Profile wizard work against the same file.
-                // Without this, the first wizard save creates a TOML
-                // containing only the new profile, silently shadowing
-                // everything in profiles.lua.
-                do {
-                    let toml = try importer.convertToTOML(lua)
-                    try FileManager.default.createDirectory(
-                        at: tomlURL.deletingLastPathComponent(),
-                        withIntermediateDirectories: true
-                    )
-                    try toml.write(to: tomlURL, atomically: true, encoding: .utf8)
-                    logger.info("migrated \(profiles.count) profiles from \(luaURL.path) (Hammerspoon) → \(tomlURL.path); future edits should go to the TOML file")
-                } catch {
-                    logger.warn("loaded \(profiles.count) profiles from \(luaURL.path) but migration to \(tomlURL.path) failed: \(error). Wizard saves won't include these until migration succeeds.")
-                }
-                return profiles
-            } catch {
-                logger.warn("failed to import \(luaURL.path): \(error)")
-            }
-        }
-
-        // Fresh user with no Hammerspoon and no Swift-app config:
-        // bootstrap a starter `profiles.toml` so the app does
-        // something useful on first launch (apply MacBook audio
-        // defaults when undocked) instead of staring blankly until
-        // the user creates a config.
-        do {
-            try writeStarterConfig(at: tomlURL)
-            let profiles = try ConfigLoader().loadProfiles(from: tomlURL)
-            logger.info("first launch — wrote a starter config to \(tomlURL.path) (\(profiles.count) profile)")
-            return profiles
-        } catch {
-            logger.warn("failed to write starter config to \(tomlURL.path): \(error) — engine will run idle until a config is created")
-            return []
-        }
-    }
-
-    /// Default `profiles.toml` content, written on first launch when
-    /// neither an existing TOML nor a Hammerspoon `profiles.lua`
-    /// exists. Includes one always-matches "laptop" profile with
-    /// modern-Mac defaults plus inline comments showing the user how
-    /// to add docked locations and OBS scenes.
-    private static let starterConfig = """
-    # AV Pain Reliever — profile config.
-    # Each [profiles.<name>] section defines a location.
-    #
-    # A profile matches when every USB device in its `fingerprint` is
-    # currently attached. Most-specific match wins; alphabetical
-    # tiebreak. An empty fingerprint matches always (specificity 0),
-    # making such a profile the implicit fallback when nothing more
-    # specific matches.
-
-    [profiles.laptop]
-    # Implicit fallback — no fingerprint means this matches whenever
-    # no docked profile does (typical state: undocked MacBook).
-    audioInput  = "MacBook Pro Microphone"
-    audioOutput = "MacBook Pro Speakers"
-
-    # Add a docked profile here, e.g.:
-    #
-    # [profiles.home-office]
-    # audioInput  = "Yeti Stereo Microphone"
-    # audioOutput = "External DAC"
-    # fingerprint = [
-    #   { vendorID = 0x2188, productID = 0x6533, name = "Dock" },
-    #   { vendorID = 0x043e, productID = 0x9a68, name = "External camera" },
-    # ]
-    #
-    # Easier: click the menu bar → Add Profile… and the wizard
-    # captures your currently-attached devices automatically.
-
-    """
-
-    private func writeStarterConfig(at url: URL) throws {
-        let parent = url.deletingLastPathComponent()
-        try FileManager.default.createDirectory(
-            at: parent, withIntermediateDirectories: true
-        )
-        try Self.starterConfig.write(to: url, atomically: true, encoding: .utf8)
-    }
 }
